requirements.txt
Flask
PyGithub
cryptography
psutil